Issue 44 removing images

+++ src/image/image.service.ts
@@ -2,6 +2,8 @@ import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model, Types } from 'mongoose';
 import { Image } from 'src/common/schemas';
+import * as fs from 'fs';
+import { ImageInterceptorEnum } from 'src/common/enums';
 
 @Injectable()
 export class ImageService {
@@ -18,8 +20,47 @@ export class ImageService {
     return image;
   }
 
-  async remove(id: Types.ObjectId) {
+  async remove(id: Types.ObjectId, subpath: ImageInterceptorEnum) {
     const image = await this.findOne(id);
-    await this.imageModel.findByIdAndDelete(image._id);
+    if (image.filename) {
+      const path = `${process.cwd()}/images/${subpath}/${image.filename}`;
+      try {
+        fs.unlinkSync(path);
+        return await this.imageModel.findByIdAndDelete(image._id);
+      } catch (err) {
+        console.error(err);
+      }
+    }
+    throw new NotFoundException('Image not Found');
+  }
+
+  async removeMany(ids: Types.ObjectId[], subpath: ImageInterceptorEnum) {
+    const images = await this.imageModel.find({ _id: { $in: ids } });
+    if (images.length) {
+      const path = `${process.cwd()}/images/${subpath}`;
+      try {
+        images.forEach((image) => {
+          fs.unlinkSync(`${path}/${image.filename}`);
+        });
+        return await this.imageModel.deleteMany({ _id: { $in: ids } });
+      } catch (err) {
+        console.error(err);
+        throw new NotFoundException('Images not Found');
+      }
+    }
+  }
+
+  // Function for removing images by filenames. Not created in DB yet.
+  async removeByFilenames(filenames: string[], subpath: ImageInterceptorEnum) {
+    const path = `${process.cwd()}/images/${subpath}`;
+
+    try {
+      return filenames.forEach((filename) =>
+        fs.unlinkSync(`${path}/${filename}`),
+      );
+    } catch (err) {
+      console.error(err);
+      throw new NotFoundException('Images not Found');
+    }
   }
 }
